Move RecordSet actions into own file
